pyproject.toml
[tool.poetry]
name = "lucyna-api"
version = "0.1.0"
description = ""
authors = ["Kiba <munyontodd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
awswrangler = "^3.9.0"
uvicorn = "^0.30.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.31"
requests = "^2.32.3"
pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
